requirements.txt
amqp==2.4.1
async-timeout==4.0.2
billiard==3.5.0.5
certifi==2018.11.29
chardet==3.0.4
Click==7.0
croniter==0.3.29
Django==2.1.7
django-crispy-forms==1.7.2
django-tinymce==2.8.0
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
feedparser==6.0.10
gunicorn==20.1.0
idna==2.8
jsonfield==2.0.2
kombu==4.3.0
python-dateutil==2.8.0
python-memcached==1.59
pytz==2018.9
redis==4.5.5
requests==2.21.0
rq==1.15.0
rq-scheduler==0.9
sgmllib3k==1.0.0
six==1.12.0
soupsieve==1.8
stripe==2.21.0
urllib3==1.24.1
vine==1.2.0
whitenoise==6.4.0
django-environ==0.10.0
bleach==6.0.0
